requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.1.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
